requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pytz>=2024.1
tabulate>=0.9.0
pytest>=7.4.0
# playwright is optional; install only if needed for JS-rendered fallback:
# playwright>=1.40.0
